+++ src/main/java/com/alexvr/bedres/gui/FluxOracleScreen.java
@@ -27,7 +27,7 @@ public class FluxOracleScreen extends ForgeIngameGui {
         super.renderGameOverlay(partialTicks);
         drawModalRectWithCustomSizedTexture(0,32,Minecraft.getInstance().mainWindow.getScaledHeight(),Minecraft.getInstance().mainWindow.getScaledHeight()-64,new ResourceLocation("bedres","textures/gui/flux_oracle_gui.png"));
         drawString(BedrockResources.proxy.getMinecraft().fontRenderer,flux.getBedrockFluxString(),1,Minecraft.getInstance().mainWindow.getScaledHeight()-72,16777215);
-        float ratio = flux.getBedrockFlux()/flux.getMaxBedrockFlux();
+        float ratio = (float) (flux.getBedrockFlux()/flux.getMaxBedrockFlux());
         mc.getTextureManager().bindTexture(ScrapeTankScreen.SCRAPE_TANK_GUI_TEXTURE);
         this.blit(5, (Minecraft.getInstance().mainWindow.getScaledHeight()-(59)), 179, 16, 42, (int)(59*ratio));
         this.blit(11, (Minecraft.getInstance().mainWindow.getScaledHeight()-(59)), 179, 16, 42, (int)(59*ratio));
